+++ Source/LowEntryExtendedStandardLibrary/Public/Classes/LowEntryHashingSha512Library.h
@@ -25,17 +25,17 @@ private:
 	int64			h6 = 0x1f83d9abfb41bd6bL;
 	int64			h7 = 0x5be0cd19137e2179L;
 
-	int32			count = 0;
+	int64			count = 0;
 	TArray<uint8>	buffer;
 
 
 private:
 	void initialize();
-	void update(const TArray<uint8>& b, const int32 offset, const int32 len);
+	void update(const TArray<uint8>& b, const int64 offset, const int64 len);
 	TArray<uint8> digest();
 	TArray<uint8> padBuffer();
 	TArray<uint8> getResult();
-	void sha(const TArray<uint8>& in, int32 offset);
+	void sha(const TArray<uint8>& in, int64 offset);
 	FORCEINLINE static int64 s(const int64 a, const int32 b);
